+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## [3.0.0] - 2023-05-28
+
+### Breaking
+
+- Options renamed to Settings
+- Settings object is required
+- At least one fallback must be specified
+- `poll` must be > 0 when specified
+
+### Added
+
+- `forceFallbacks` setting
+
 ## [2.0.0] - 2023-05-28
 
 ### Breaking